code/requirements.txt
amqp==2.6.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==3.6.4.0
black==24.1.1
celery==4.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.0.10
djangorestframework==3.14.0
kombu==4.6.11
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
vine==1.3.0
wcwidth==0.2.13
gunicorn==21.2.0
